+++ openpyxl/reader/tests/test_style.py
@@ -10,6 +10,7 @@ from zipfile import ZipFile
 from openpyxl.compat import safe_string, OrderedDict
 from openpyxl.reader.excel import load_workbook
 from openpyxl.utils.indexed_list import IndexedList
+from openpyxl.xml.functions import fromstring
 
 from openpyxl.styles import (
     borders,
@@ -267,3 +268,42 @@ def test_rgb_colors(StyleReader, datadir):
     assert len(reader.color_index) == 64
     assert reader.color_index[0] == "00000000"
     assert reader.color_index[-1] == "00333333"
+
+
+def test_custom_number_formats(StyleReader, datadir):
+    datadir.chdir()
+    with open("styles_number_formats.xml") as src:
+        reader = StyleReader(src.read())
+
+    reader.parse_custom_num_formats()
+    assert reader.custom_number_formats == {
+        43:'_ * #,##0.00_ ;_ * \-#,##0.00_ ;_ * "-"??_ ;_ @_ ',
+        176: "#,##0.00_ ",
+        180: "yyyy/m/d;@",
+        181: "0.00000_ "
+    }
+    assert reader.number_formats == [
+        '_ * #,##0.00_ ;_ * \-#,##0.00_ ;_ * "-"??_ ;_ @_ ',
+        "#,##0.00_ ",
+        "yyyy/m/d;@",
+        "0.00000_ "
+    ]
+
+
+def test_assign_number_formats(StyleReader):
+
+    reader = StyleReader("<root />")
+    reader.custom_number_formats = {43:'_ * #,##0.00_ ;_ * \-#,##0.00_ ;_ * "-"??_ ;_ @_ '}
+    reader.number_formats = IndexedList(['_ * #,##0.00_ ;_ * \-#,##0.00_ ;_ * "-"??_ ;_ @_ '])
+
+    node = fromstring("""
+    <xf xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"
+         numFmtId="43" fontId="2" fillId="0" borderId="0"
+         applyFont="0" applyFill="0" applyBorder="0" applyAlignment="0" applyProtection="0">
+          <alignment vertical="center"/>
+    </xf>
+    """)
+    reader._parse_xfs(node)
+
+    assert reader.cell_styles == [{'fillId': 0, 'fontId': 2, 'xfId': 0,
+                                   'protectionId': 0, 'borderId': 0, 'numFmtId': 164, 'alignmentId': 0}]