+++ openpyxl/styles/styleable.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
-from collections import namedtuple
+from warnings import warn
 
 from openpyxl.utils.indexed_list import IndexedList
 from .numbers import BUILTIN_FORMATS, BUILTIN_FORMATS_REVERSE
@@ -62,9 +62,11 @@ class StyleableObject(object):
     alignment = StyleDescriptor('_alignments', '_alignment_id')
 
     __slots__ = ('parent', '_font_id', '_border_id', '_fill_id',
-                 '_alignment_id', '_protection_id', '_number_format_id')
+                 '_alignment_id', '_protection_id', '_number_format_id', 'pivotButton',
+                 'quotePrefix')
 
-    def __init__(self, sheet, fontId=0, fillId=0, borderId=0, alignmentId=0, protectionId=0, numFmtId=0):
+    def __init__(self, sheet, fontId=0, fillId=0, borderId=0, alignmentId=0,
+                 protectionId=0, numFmtId=0, pivotButton=None, quotePrefix=None):
         self._font_id = fontId
         self._fill_id = fillId
         self._border_id = borderId
@@ -72,6 +74,8 @@ class StyleableObject(object):
         self._protection_id = protectionId
         self._number_format_id = numFmtId
         self.parent = sheet
+        self.pivotButton = pivotButton
+        self.quotePrefix = quotePrefix
 
 
     @property
@@ -82,7 +86,10 @@ class StyleableObject(object):
             fillId=self._fill_id,
             fontId=self._font_id,
             numFmtId=self._number_format_id,
-            protectionId=self._protection_id)
+            protectionId=self._protection_id,
+            pivotButton=self.pivotButton,
+            quotePrefix=self.quotePrefix
+        )
 
         return self.parent.parent._cell_styles.add(style)
 
@@ -93,26 +100,30 @@ class StyleableObject(object):
                or self._fill_id
                or self._font_id
                or self._number_format_id
-               or self._protection_id)
+               or self._protection_id
+               or self.pivotButton
+               or self.quotePrefix
+               )
 
     #legacy
     @property
     def style(self):
+        warn("Use formatting objects such as font directly")
         return Style(
-            font=self.font,
-            fill=self.fill,
-            border=self.border,
-            alignment=self.alignment,
+            font=self.font.copy(),
+            fill=self.fill.copy(),
+            border=self.border.copy(),
+            alignment=self.alignment.copy(),
             number_format=self.number_format,
-            protection=self.protection
+            protection=self.protection.copy()
         )
 
     #legacy
     @style.setter
     def style(self, value):
-        self.font = value.font._StyleProxy__target
-        self.fill = value.fill._StyleProxy__target
-        self.border = value.border._StyleProxy__target
-        self.protection = value.protection._StyleProxy__target
-        self.alignment = value.alignment._StyleProxy__target
+        self.font = value.font.copy()
+        self.fill = value.fill.copy()
+        self.border = value.border.copy()
+        self.protection = value.protection.copy()
+        self.alignment = value.alignment.copy()
         self.number_format = value.number_format