+++ openpyxl/worksheet/header_footer.py
@@ -109,7 +109,7 @@ class HeaderFooterItem(object):
 
         m = SIZE_REGEX.search(text)
         if m:
-            self.font_size = m.group(0)
+            self.font_size = int(m.group('size'))
             text = SIZE_REGEX.sub('', text)
 
         m = COLOR_REGEX.search(text)
@@ -194,14 +194,23 @@ ITEM_REGEX = re.compile("""
 (&L(?P<left>.+?))?
 (&C(?P<center>.+?))?
 (&R(?P<right>.+?))?
-$""", re.VERBOSE)
+$""", re.VERBOSE | re.DOTALL)
+
+# add support for multiline strings (how do re.flags combine?)
+
+from warnings import warn
 
 def _split_string(text):
     """Split the combined (decoded) string into left, center and right parts"""
     m = ITEM_REGEX.match(text)
-    return m.groupdict()
+    try:
+        parts = m.groupdict()
+    except AttributeError:
+        warn("""Cannot parse header or footer so it will be ignored""")
+        parts = {'left':'', 'right':'', 'center':''}
+    return parts
 
 HEADER_REGEX = re.compile(r"(&[ABDEGHINOPSTUXYZ\+\-])") # split part into commands
 FONT_REGEX = re.compile('&"(?P<font>.+)"')
 COLOR_REGEX = re.compile("&K(?P<color>[A-F0-9]{6})")
-SIZE_REGEX = re.compile(r"&\d+")
+SIZE_REGEX = re.compile(r"&(?P<size>\d+)")