+++ openpyxl/writer/relations.py
@@ -11,7 +11,7 @@ from openpyxl.xml.constants import (
 )
 
 
-def write_rels(worksheet, drawing_id, comments_id):
+def write_rels(worksheet, drawing_id, comments_id, vba_controls_id):
     """Write relationships for the worksheet to xml."""
     root = Element('{%s}Relationships' % PKG_REL_NS)
     for rel in worksheet.relationships:
@@ -35,5 +35,9 @@ def write_rels(worksheet, drawing_id, comments_id):
                  'Type': VML_NS,
                  'Target': '../drawings/commentsDrawing%s.vml' % comments_id}
         SubElement(root, '{%s}Relationship' % PKG_REL_NS, attrs)
+    if worksheet.vba_controls is not None:
+        attrs = {'Id': worksheet.vba_controls,
+                 'Type': VML_NS,
+                 'Target': '../drawings/vmlDrawing%s.vml' % vba_controls_id}
+        SubElement(root, '{%s}Relationship' % PKG_REL_NS, attrs)
     return root
-